requirements.txt
python-dotenv==0.21.1
langchain==0.0.101
openai==0.27.2
wolframalpha==5.0.0
transformers==4.26.1
torch==2.0.0
discord.py==2.1.0
revChatGPT==3.2.0
faiss-cpu==1.7.3
nltk==3.6.5
numpy==1.24.2